chore: improvement traverse DOM on keyboard delete event

+++ src/lib/core/GeoEditor.ts
@@ -3337,7 +3337,7 @@ export class GeoEditor implements IControl {
         e.preventDefault();
       }
       // Delete
-      if ( (e.key === 'Delete' || e.key === 'Backspace') && !isInputField) {
+      if ((e.key === 'Delete' || e.key === 'Backspace') && !isInputField && e.target instanceof Element && e.target.closest(`.${CSS_PREFIX}-attribute-panel`)) {
         this.deleteSelectedFeatures();
         e.preventDefault();
       }
